backend/requirements.txt
# ── Core ──────────────────────────────────────────────────────
Flask==3.1.0
flask-cors==5.0.0
gunicorn==22.0.0

# ── MySQL connector (pure Python, no C extension needed) ───────
mysql-connector-python==9.1.0

# ── Env / Config ──────────────────────────────────────────────
python-dotenv==1.0.1

# ── Production extras ─────────────────────────────────────────
# Structured JSON logging for ELK:
# python-json-logger==2.0.7